Add files via upload
for the subsystem, I added 3 left and right motors, then group them, then made a up, down, left, and right function. Then for robotcontainer, I made an xboxcontroller object then made two method which get the x and y of the joystick. for command I made a movewithcontroller method, which will see the input for leftx and left y and move the motor. I think we invert motor because the left and right motor needs to be the opposite.